Moved data managers / generators to new directory.
