+++ app/src/androidTest/java/com/koleff/kare_android/statistics/ExerciseStatisticsUseCaseTest.kt
@@ -5,7 +5,7 @@ import androidx.room.Room
 import androidx.test.core.app.ApplicationProvider
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.google.firebase.crashlytics.internal.Logger
-import com.koleff.kare_android.common.MockupDataGeneratorV2
+import com.koleff.kare_android.common.manager.data.MockupDataGeneratorV2
 import com.koleff.kare_android.data.datasource.DoWorkoutPerformanceMetricsLocalDataSource
 import com.koleff.kare_android.data.datasource.ExerciseStatisticsDataSource
 import com.koleff.kare_android.data.datasource.ExerciseStatisticsLocalDataSource
