+++ app/src/main/java/com/koleff/kare_android/data/room/manager/WorkoutDBManager.kt
@@ -1,8 +1,8 @@
 //package com.koleff.kare_android.data.room.manager
 //
-//import com.koleff.kare_android.common.ExerciseGenerator
-//import com.koleff.kare_android.common.MockupDataGeneratorV2
-//import com.koleff.kare_android.common.WorkoutGenerator
+//import com.koleff.kare_android.common.manager.data.ExerciseGenerator
+//import com.koleff.kare_android.common.manager.data.MockupDataGeneratorV2
+//import com.koleff.kare_android.common.manager.data.WorkoutGenerator
 //import com.koleff.kare_android.data.room.dao.ExerciseDao
 //import com.koleff.kare_android.data.room.dao.ExerciseSetDao
 //import com.koleff.kare_android.data.room.dao.WorkoutDao
