+++ app/src/main/java/com/koleff/kare_android/ui/compose/banners/ExerciseBanner.kt
@@ -60,7 +60,7 @@
 //import androidx.compose.ui.unit.dp
 //import androidx.compose.ui.unit.sp
 //import com.koleff.kare_android.R
-//import com.koleff.kare_android.common.MockupDataGeneratorV2
+//import com.koleff.kare_android.common.manager.data.MockupDataGeneratorV2
 //import com.koleff.kare_android.data.model.dto.ExerciseDto
 //import com.koleff.kare_android.data.model.dto.MuscleGroup
 //import kotlin.math.roundToInt
