+++ app/src/main/java/com/koleff/kare_android/ui/compose/components/SearchWorkoutsComponents.kt
@@ -9,7 +9,7 @@
 //import androidx.compose.ui.Modifier
 //import androidx.compose.ui.tooling.preview.Preview
 //import androidx.compose.ui.unit.dp
-//import com.koleff.kare_android.common.MockupDataGeneratorV2
+//import com.koleff.kare_android.common.manager.data.MockupDataGeneratorV2
 //import com.koleff.kare_android.data.model.dto.WorkoutDto
 //import com.koleff.kare_android.ui.compose.banners.WorkoutBanner
 //
