+++ app/src/main/java/com/koleff/kare_android/ui/compose/dialogs/FavoriteWorkoutDialog.kt
@@ -3,7 +3,7 @@ package com.koleff.kare_android.ui.compose.dialogs
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.tooling.preview.PreviewLightDark
-import com.koleff.kare_android.common.MockupDataGeneratorV2
+import com.koleff.kare_android.common.manager.data.MockupDataGeneratorV2
 import java.util.Locale
 
 @Composable
