+++ app/src/main/java/com/koleff/kare_android/ui/compose/screen/DoWorkoutScreen.kt
@@ -59,7 +59,7 @@
 //import androidx.compose.ui.unit.sp
 //import androidx.hilt.navigation.compose.hiltViewModel
 //import com.koleff.kare_android.R
-//import com.koleff.kare_android.common.MockupDataGeneratorV2
+//import com.koleff.kare_android.common.manager.data.MockupDataGeneratorV2
 //import com.koleff.kare_android.common.timer.TimerUtil
 //import com.koleff.kare_android.data.model.dto.ExerciseDto
 //import com.koleff.kare_android.data.model.dto.ExerciseProgressDto
