+++ app/src/test/java/com/koleff/kare_android/do_workout/DoWorkoutUseCasesUnitTest.kt
@@ -1,6 +1,6 @@
 package com.koleff.kare_android.do_workout
 
-import com.koleff.kare_android.common.MockupDataGeneratorV2
+import com.koleff.kare_android.common.manager.data.MockupDataGeneratorV2
 import com.koleff.kare_android.data.datasource.DoWorkoutLocalDataSource
 import com.koleff.kare_android.data.model.dto.ExerciseDto
 import com.koleff.kare_android.data.model.dto.ExerciseSetDto
